# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CertEfficiency API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent if account exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password using token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password for authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /user-certifications:
    get:
      operationId: listUserCertifications
      tags: [certifications]
      summary: List user's selected certifications
      responses:
        "200":
          description: List of user certifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserCertification"
    post:
      operationId: createUserCertification
      tags: [certifications]
      summary: Add a certification for the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCertificationInput"
      responses:
        "201":
          description: User certification created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCertification"
  /user-certifications/{id}:
    patch:
      operationId: updateUserCertification
      tags: [certifications]
      summary: Update user certification details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCertificationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCertification"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Auth schemas
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        password:
          type: string
          minLength: 8
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    ResetPasswordInput:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
    User:
      type: object
      required: [id, name, email, plan, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        plan:
          type: string
          enum: [guest, free, pro]
        createdAt:
          type: string
          format: date-time
    AuthResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # Certification schemas
    Certification:
      type: object
      required: [id, name, code, category]
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
    UserCertification:
      type: object
      required: [id, userId, certificationId, certification, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        certificationId:
          type: string
        certification:
          $ref: "#/components/schemas/Certification"
        examDate:
          type: ["string", "null"]
          format: date
        weeklyHours:
          type: ["integer", "null"]
        confidenceLevel:
          type: ["string", "null"]
          enum: [beginner, intermediate, advanced, null]
        attemptedBefore:
          type: ["boolean", "null"]
        preferredStyle:
          type: ["string", "null"]
        isPrimary:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserCertificationInput:
      type: object
      required: [certificationId]
      properties:
        certificationId:
          type: string
        examDate:
          type: string
          format: date
        weeklyHours:
          type: integer
        confidenceLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        attemptedBefore:
          type: boolean
        preferredStyle:
          type: string
        isPrimary:
          type: boolean
    UserCertificationUpdate:
      type: object
      properties:
        examDate:
          type: string
          format: date
        weeklyHours:
          type: integer
        confidenceLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        attemptedBefore:
          type: boolean
        preferredStyle:
          type: string
        isPrimary:
          type: boolean

    # Conversation schemas
...